+++ src/main/java/com/drumond/rentalcar/models/Rent.java
@@ -2,8 +2,6 @@ package com.drumond.rentalcar.models;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
 import java.time.ZonedDateTime;
@@ -15,7 +13,7 @@ public class Rent {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User client;
